Move common types
